Force MSIX table to use BRAM.
For some reason Vivado can't infer BRAMs for some of the Migen Memory
primitives.  You seem to need to use the READ_FIRST mode of the ports.

This commit also reduces the size of the table to 16 entries from the
default 2k.  The ACS tests only seem to use a single entry for testing
so there's no need to have a 2k vector table.