Fix bug in ATS completion routing
This commit fixes a bug in the ATS completion routing path from the
depacketizer.  The crossbar routes master completions based on channel
ID and the ATS engine used the same ID as the DMA.  The solution is to
use the port.channel attribute returned when calling
crossbar.get_master_port().