Update TXN monitor with overflow detection
This commit adds fields to the BSA Transaction Monitor Control Register
to make it a slighly nicer target for software.  The BSA tests should
never issue more TLP requests than there are FIFO entries in the
transaction monitor.  However, with a single Enable bit in the control
register, it makes it hard to test functional behaviour in verification.
To fix this, clear/flush and overflow bits are added to the control
register that allow software to detect overflows and/or flush the FIFO
to clean out stale TLPs.

This commit also adds tests that verify overflow behaviour and make sure
the hardware can recover following an overflow event.