Timing fix
This commit fixes a long combinational path in the ATC that prevented
the design meeting timing.  The ATS translation end address is computed
when the ATS response is received and stored in the ATC.  The end
address is used in the invalidation logic which also replicated the same
arithmetic operations.